初步完成“上次收款信息”和数量修改为double

+++ LocalERP/DataAccess/Data/SellProfit.cs
@@ -8,7 +8,7 @@ namespace LocalERP.DataAccess.Data
     public class SellProfit
     {
         public int ID;
-        public int cnt;
+        public double cnt;
         public double price;
         public double sum_price;
         public double cost;
